requirements.txt
openai-whisper
torch
requests
python-dotenv
